finish demonstration of propgating null for non present columns

+++ code/flink-sql/01-process-null/cc-flink/ddl.src_tickets.sql
@@ -1,7 +1,7 @@
 CREATE TABLE src_tickets (
     case_id STRING NOT NULL,
     description STRING,
-    priority INT,
+    priority BIGINT,
     owner STRING,
     testresults STRING,
     creation_ts TIMESTAMP_LTZ(3),
